requirements.txt

djangorestframework  
django-allauth  
django-crispy-forms  
django-debug-toolbar  
channels  
celery  
django-storages  
django-filter  
requests  
pandas  
numpy  
matplotlib  
Flask  
beautifulsoup4  
SQLAlchemy   
tweepy  
stripe  
boto3  
twilio